fix: move to components, working types
