+++ apps/contribute/components/Profile/index.tsx
@@ -5,8 +5,8 @@ import styled from 'styled-components';
 
 import { COLOR, MEDIA_QUERY, NA, areAddressesEqual } from '@autonolas/frontend-library';
 
-import { BadgeLoading, ShowBadge } from 'common-util/ShowBadge';
-import TruncatedEthereumLink from 'common-util/TruncatedEthereumLink';
+import { BadgeLoading, ShowBadge } from 'components/ShowBadge';
+import TruncatedEthereumLink from 'components/TruncatedEthereumLink';
 import { getLatestMintedNft, updateUserStakingData } from 'common-util/api';
 import { getName, getTier } from 'common-util/functions';
 import { useAppSelector } from 'store/setup';
