+++ apps/contribute/pages/chatbot.jsx
@@ -1,6 +1,6 @@
 // import Chatbot from 'components/Chatbot';
 // export default Chatbot;
 // TODO: It will be added once the page is updated.
-import { PageDoesNotExist } from 'common-util/PageDoesNotExist';
+import { PageDoesNotExist } from 'components/PageDoesNotExist';
 
 export default PageDoesNotExist;
