+++ apps/contribute/pages/members.jsx
@@ -1,6 +1,6 @@
 // import Members from 'components/Members';
 // export default Members;
 // TODO: It will be added once the page is updated.
-import { PageDoesNotExist } from 'common-util/PageDoesNotExist';
+import { PageDoesNotExist } from 'components/PageDoesNotExist';
 
 export default PageDoesNotExist;
