+++ apps/contribute/pages/predict.jsx
@@ -1,6 +1,6 @@
 // import Predict from 'components/Predict';
 // export default Predict;
 // TODO: It will be added once the page is updated.
-import { PageDoesNotExist } from 'common-util/PageDoesNotExist';
+import { PageDoesNotExist } from 'components/PageDoesNotExist';
 
 export default PageDoesNotExist;
