+++ apps/contribute/pages/verification.jsx
@@ -1,6 +1,6 @@
 // import Verification from 'components/Verification';
 // export default Verification;
 // TODO: It will be added once the page is updated.
-import { PageDoesNotExist } from 'common-util/PageDoesNotExist';
+import { PageDoesNotExist } from 'components/PageDoesNotExist';
 
 export default PageDoesNotExist;
