+++ tsconfig.base.json
@@ -34,7 +34,7 @@
         "libs/common-middleware/src/index.ts"
       ],
       "@autonolas-frontend-mono/common-middleware/server": [
-        "libs/common-middleware/src/server.ts"
+        "libs/common-middleware/src/middleware.ts"
       ],
       "@autonolas-frontend-mono/feature-service-status-info": [
         "libs/feature-service-status-info/src/index.ts"
